+++ src/integrations/integr_pdb.rs
@@ -17,7 +17,7 @@ use crate::call_validation::{ContextEnum, ChatMessage};
 use crate::integrations::sessions::{IntegrationSession, get_session_hashmap_key};
 use crate::global_context::GlobalContext;
 use crate::tools::tools_description::Tool;
-use crate::integrations::io_helper::{first_n_chars, last_n_chars, last_n_lines, write_to_stdin_and_flush, read_until_token_or_timeout};
+use crate::integrations::process_io_utils::{first_n_chars, last_n_chars, last_n_lines, write_to_stdin_and_flush, read_until_token_or_timeout};
 
 const SESSION_TIMEOUT_AFTER_INACTIVITY: Duration = Duration::from_secs(30 * 60);
 const PDB_TOKEN: &str = "(Pdb)";
