+++ src/integrations/mod.rs
@@ -1,7 +1,7 @@
 pub mod integr_github;
 pub mod integr_pdb;
 pub mod sessions;
-pub mod io_helper;
+pub mod process_io_utils;
 
 pub const INTEGRATIONS_DEFAULT_YAML: &str = r#"# This file is used to configure integrations in Refact Agent.
 # If there is a syntax error in this file, no integrations will work.
